English v0.1
First release of the English Translation.
There are still many aspects of the UI to translate:
Song names are translated, but not with Official Titles.
Lyrics remain in Japanese, not English or Romaji.
Entry screen is mostly translated and can likely be navigated with no knowledge of Japanese.
Supports Lucid Dreaming and Out of Eden song titles.